separate out validation tests
